n8n_template_manager/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
requests
python-whoosh
sqlite
